+++ test/unit/iam/utils.js
@@ -22,6 +22,9 @@ const PASSPHRASE = require('./constants').PASSPHRASE;
 const ST_HEADER = require('./constants').ST_HEADER;
 const ST_SIG = require('./constants').ST_SIG;
 const TEST_DIR = require('./constants').TEST_DIR;
+const DEFAULT_OCI_DIR = require('./constants').DEFAULT_OCI_DIR;
+const DEFAULT_OCI_FILE = require('./constants').DEFAULT_OCI_FILE;
+const DEFAULT_OCI_FILE_BACKUP = require('./constants').DEFAULT_OCI_FILE_BACKUP;
 
 const AUTH_HEADER_PATTERN = new RegExp('^Signature headers=".+?",\
 keyId="(.+?)",algorithm="(.+?)",signature="(.+?)",version="(.+?)"$');
@@ -246,6 +249,32 @@ function writeFileLines(file, lines) {
     fs.writeFileSync(file, lines.join('\n'));
 }
 
+function backupDefaultOCIFile() {
+    try {
+        if (!fs.existsSync(DEFAULT_OCI_DIR)) {
+            fs.mkdirSync(DEFAULT_OCI_DIR);
+        }
+        if (fs.existsSync(DEFAULT_OCI_FILE)) {
+            fs.copyFileSync(DEFAULT_OCI_FILE, DEFAULT_OCI_FILE_BACKUP);
+        }
+    } catch(err) {
+        throw new Error(`This test uses directory ${DEFAULT_OCI_DIR} and \
+file ${DEFAULT_OCI_FILE}.  Please make sure they either exist or can be \
+created.  Error: ${err.message}`);
+    }
+}
+
+//only call after backupDefaultOCIFile() successfully completed
+function restoreDefaultOCIFile() {
+    if (fs.existsSync(DEFAULT_OCI_FILE_BACKUP)) {
+        fs.renameSync(DEFAULT_OCI_FILE_BACKUP, DEFAULT_OCI_FILE);
+    } else if (fs.existsSync(DEFAULT_OCI_FILE)) {
+        //In case the config did not exist before the test, we remove the one
+        //created by the test.
+        fs.unlinkSync(DEFAULT_OCI_FILE);
+    }
+}
+
 module.exports = {
     inspect,
     iam2cfg,
@@ -260,5 +289,7 @@ module.exports = {
     verifyAuthLaterDate,
     makeTestDir,
     removeTestDir,
-    writeFileLines
+    writeFileLines,
+    backupDefaultOCIFile,
+    restoreDefaultOCIFile
 };